Update .gitignore and refactor cactus.cpp
- Added ios/cactus.xcframework/ and android/src/main/jniLibs/ to .gitignore to prevent tracking of build artifacts.

- Refactored cactus.cpp to simplify the file structure by including all module implementations directly, enhancing maintainability and readability.
